IAM - add tests for new OIDC functionality
